+++ src/main/java/com/example/freshinsights/repository/ActivityRepository.java
@@ -4,8 +4,10 @@ import com.example.freshinsights.model.Activity;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface ActivityRepository extends CrudRepository<Activity, Integer>
+public interface ActivityRepository extends CrudRepository<Activity, BigInteger>
 {
 
 }
