+++ src/main/java/com/example/freshinsights/repository/EmployeeRepository.java
@@ -4,8 +4,10 @@ import com.example.freshinsights.model.Employee;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface EmployeeRepository extends CrudRepository<Employee, Integer>
+public interface EmployeeRepository extends CrudRepository<Employee, BigInteger>
 {
 
 }
