+++ src/main/java/com/example/freshinsights/repository/FlowRepository.java
@@ -4,8 +4,10 @@ import com.example.freshinsights.model.Flow;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface FlowRepository extends CrudRepository<Flow, Integer>
+public interface FlowRepository extends CrudRepository<Flow, BigInteger>
 {
 
 }
