+++ src/main/java/com/example/freshinsights/repository/FlowStepsRepository.java
@@ -4,8 +4,10 @@ import com.example.freshinsights.model.FlowSteps;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface FlowStepsRepository extends CrudRepository<FlowSteps, Integer>
+public interface FlowStepsRepository extends CrudRepository<FlowSteps, BigInteger>
 {
 
 }
