+++ src/main/java/com/example/freshinsights/repository/ProductsRepository.java
@@ -4,8 +4,10 @@ import com.example.freshinsights.model.Products;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface ProductsRepository extends CrudRepository<Products, Integer>
+public interface ProductsRepository extends CrudRepository<Products, BigInteger>
 {
 
 }
