+++ src/main/java/com/example/freshinsights/service/FlowService.java
@@ -5,6 +5,7 @@ import com.example.freshinsights.repository.FlowRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -16,13 +17,13 @@ public class FlowService
     public void createFlow(Flow flow) {
         flowRepository.save(flow);}
 
-    public void deleteFlow(int id) {
+    public void deleteFlow(BigInteger id) {
         flowRepository.deleteById(id);
     }
 
     public List<Flow> findAllFlows() {return (List<Flow>) flowRepository.findAll();
     }
 
-    public Flow findFlowByID(int id) {return flowRepository.findById(id).get();
+    public Flow findFlowByID(BigInteger id) {return flowRepository.findById(id).get();
     }
 }
