+++ src/main/java/com/example/freshinsights/service/ProductsService.java
@@ -6,6 +6,7 @@ import com.example.freshinsights.repository.ProductsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -16,12 +17,12 @@ public class ProductsService
 
     public void createProduct(Products products) {productsRepository.save(products);}
 
-    public void deleteProduct(int id) {productsRepository.deleteById(id);
+    public void deleteProduct(BigInteger id) {productsRepository.deleteById(id);
     }
 
     public List<Products> findAllProducts() {return (List<Products>) productsRepository.findAll();
     }
 
-    public Products findProductByID(int id) {return productsRepository.findById(id).get();
+    public Products findProductByID(BigInteger id) {return productsRepository.findById(id).get();
     }
 }
